moved detectors to separate folder, updated graysacle test
