+++ NVs.OccupancySensor.CV/Impl/Detectors/HOG/HOGDescriptorWrapper.cs
@@ -1,7 +1,7 @@
 using Emgu.CV;
 using Emgu.CV.Structure;
 
-namespace NVs.OccupancySensor.CV.Impl.HOG
+namespace NVs.OccupancySensor.CV.Impl.Detectors.HOG
 {
     sealed class HOGDescriptorWrapper : IHOGDescriptorWrapper
     {
