+++ NVs.OccupancySensor.CV/Impl/Detectors/HOG/IHOGDescriptorWrapper.cs
@@ -2,7 +2,7 @@ using System;
 using Emgu.CV;
 using Emgu.CV.Structure;
 
-namespace NVs.OccupancySensor.CV.Impl.HOG
+namespace NVs.OccupancySensor.CV.Impl.Detectors.HOG
 {
     interface IHOGDescriptorWrapper : IDisposable 
     {
