+++ NVs.OccupancySensor.CV/Impl/Detectors/PeopleDetectorBase.cs
@@ -8,7 +8,7 @@ using Emgu.CV.Structure;
 using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 
-namespace NVs.OccupancySensor.CV.Impl
+namespace NVs.OccupancySensor.CV.Impl.Detectors
 {
     internal abstract class PeopleDetectorBase : IPeopleDetector, IDisposable
     {
